+++ zaifapi/impl.py
@@ -1,211 +1,4 @@
 # -*- coding: utf-8 -*-
-import json
-import hmac
-import hashlib
-import inspect
-from websocket import create_connection
-from zaifapi.public import ZaifPublicApiBase, FuturesPublicApiValidator
-from zaifapi.trade import ZaifTradeApiBase
-from zaifapi.url import ApiUrl
-
 
 _MAX_COUNT = 1000
 _MIN_WAIT_TIME_SEC = 1
-
-
-def _method_name():
-    return inspect.stack()[1][3]
-
-
-class ZaifPublicApi(ZaifPublicApiBase):
-    def __init__(self):
-        super().__init__(
-            ApiUrl(api_name='api', version=1)
-        )
-
-    def last_price(self, currency_pair):
-        schema_keys = ['currency_pair']
-        return self._execute_api(_method_name(), schema_keys, currency_pair=currency_pair)
-
-    def ticker(self, currency_pair):
-        schema_keys = ['currency_pair']
-        return self._execute_api(_method_name(), schema_keys, currency_pair=currency_pair)
-
-    def trades(self, currency_pair):
-        schema_keys = ['currency_pair']
-        return self._execute_api(_method_name(), schema_keys, currency_pair=currency_pair)
-
-    def depth(self, currency_pair):
-        schema_keys = ['currency_pair']
-        return self._execute_api(_method_name(), schema_keys, currency_pair=currency_pair)
-
-    def currency_pairs(self, currency_pair):
-        schema_keys = ['currency_pair']
-        return self._execute_api(_method_name(), schema_keys, currency_pair=currency_pair)
-
-    def currencies(self, currency):
-        schema_keys = ['currency']
-        return self._execute_api(_method_name(), schema_keys, currency=currency)
-
-
-class ZaifFuturesPublicApi(ZaifPublicApiBase):
-    def __init__(self):
-        super().__init__(
-            ApiUrl(api_name='fapi', version=1),
-            FuturesPublicApiValidator()
-        )
-
-    def last_price(self, group_id, currency_pair=None):
-        schema_keys = ['currency_pair', 'group_id']
-        return self._execute_api(_method_name(), schema_keys, group_id=group_id, currency_pair=currency_pair)
-
-    def ticker(self, group_id, currency_pair):
-        schema_keys = ['currency_pair', 'group_id']
-        return self._execute_api(_method_name(), schema_keys, group_id=group_id, currency_pair=currency_pair)
-
-    def trades(self, group_id, currency_pair):
-        schema_keys = ['currency_pair', 'group_id']
-        return self._execute_api(_method_name(), schema_keys, group_id=group_id, currency_pair=currency_pair)
-
-    def depth(self, group_id, currency_pair):
-        schema_keys = ['currency_pair', 'group_id']
-        return self._execute_api(_method_name(), schema_keys, group_id=group_id, currency_pair=currency_pair)
-
-    def groups(self, group_id):
-        schema_keys = ['group_id']
-        return self._execute_api(_method_name(), schema_keys, group_id=group_id)
-
-
-class ZaifPublicStreamApi(ZaifPublicApiBase):
-    def __init__(self):
-        super().__init__(
-            ApiUrl(api_name='stream',
-                   scheme='wss',
-                   host='ws.zaif.jp',
-                   port=8888)
-        )
-        self._continue = True
-
-    def stop(self):
-        self._continue = False
-
-    def execute(self, currency_pair):
-        self.params_pre_processing(['currency_pair'], params={'currency_pair': currency_pair})
-        self._url.add_param('currency_pair', currency_pair)
-        ws = create_connection(self._url.full_url())
-        while self._continue:
-            result = ws.recv()
-            yield json.loads(result)
-        ws.close()
-
-
-class ZaifTradeApi(ZaifTradeApiBase):
-    def __init__(self, key, secret):
-        super().__init__(
-            ApiUrl(api_name='tapi')
-        )
-        self._key = key
-        self._secret = secret
-
-    def get_header(self, params):
-        signature = hmac.new(bytearray(self._secret.encode('utf-8')), digestmod=hashlib.sha512)
-        signature.update(params.encode('utf-8'))
-        return {
-            'key': self._key,
-            'sign': signature.hexdigest()
-        }
-
-    def get_info(self):
-        return self._execute_api(_method_name())
-
-    def get_info2(self):
-        return self._execute_api(_method_name())
-
-    def get_personal_info(self):
-        return self._execute_api(_method_name())
-
-    def get_id_info(self):
-        return self._execute_api(_method_name())
-
-    def trade_history(self, **kwargs):
-        schema_keys = ['from_num', 'count', 'from_id', 'end_id', 'order', 'since', 'end', 'currency_pair', 'is_token']
-        return self._execute_api(_method_name(), schema_keys, kwargs)
-
-    def active_orders(self, **kwargs):
-        schema_keys = ['currency_pair', 'is_token', 'is_token_both']
-        return self._execute_api(_method_name(), schema_keys, kwargs)
-
-    def _inner_history_api(self, func_name, kwargs):
-        schema_keys = ['currency', 'from_num', 'count', 'from_id', 'end_id', 'order', 'since', 'end', 'is_token']
-        return self._execute_api(func_name, schema_keys, kwargs)
-
-    def withdraw_history(self, **kwargs):
-        return self._inner_history_api(_method_name(), kwargs)
-
-    def deposit_history(self, **kwargs):
-        return self._inner_history_api(_method_name(), kwargs)
-
-    def withdraw(self, **kwargs):
-        schema_keys = ['currency', 'address', 'message', 'amount', 'opt_fee']
-        return self._execute_api(_method_name(), schema_keys, kwargs)
-
-    def cancel_order(self, **kwargs):
-        schema_keys = ['order_id', 'is_token', 'currency_pair']
-        return self._execute_api(_method_name(), schema_keys, kwargs)
-
-    def trade(self, **kwargs):
-        schema_keys = ['currency_pair', 'action', 'price', 'amount', 'limit', 'comment']
-        return self._execute_api(_method_name(), schema_keys, kwargs)
-
-
-class ZaifTokenTradeApi(ZaifTradeApi):
-    def __init__(self, token):
-        self._token = token
-        super().__init__(None, None)
-
-    def get_header(self, params):
-        return {
-            'token': self._token
-        }
-
-
-class ZaifLeverageTradeApi(ZaifTradeApiBase):
-    def __init__(self, key, secret):
-        super().__init__(
-            ApiUrl(api_name='tlapi')
-        )
-        self._key = key
-        self._secret = secret
-
-    def get_header(self, params):
-        signature = hmac.new(bytearray(self._secret.encode('utf-8')), digestmod=hashlib.sha512)
-        signature.update(params.encode('utf-8'))
-        return {
-            'key': self._key,
-            'sign': signature.hexdigest()
-        }
-
-    def get_positions(self, **kwargs):
-        schema_keys = ['type', 'group_id', 'from_num', 'count',
-                       'from_id', 'end_id', 'order', 'since', 'end', 'currency_pair']
-        return self._execute_api(_method_name(), schema_keys, kwargs)
-
-    def position_history(self, **kwargs):
-        schema_keys = ['type', 'group_id', 'leverage_id']
-        return self._execute_api(_method_name(), schema_keys, kwargs)
-
-    def active_positions(self, **kwargs):
-        schema_keys = ['type', 'group_id', 'currency_pair']
-        return self._execute_api(_method_name(), schema_keys, kwargs)
-
-    def create_position(self, **kwargs):
-        schema_keys = ['type', 'group_id', 'currency_pair', 'action', 'price', 'amount', 'leverage', 'limit', 'stop']
-        return self._execute_api(_method_name(), schema_keys, kwargs)
-
-    def change_position(self, **kwargs):
-        schema_keys = ['type', 'group_id', 'leverage_id', 'price', 'limit', 'stop']
-        return self._execute_api(_method_name(), schema_keys, kwargs)
-
-    def cancel_position(self, **kwargs):
-        schema_keys = ['type', 'group_id', 'leverage_id']
-        return self._execute_api(_method_name(), schema_keys, kwargs)
